Code clean-up for components
