+++ front-end/src/components/AuthForm/AuthForm.css
@@ -1,0 +1,94 @@
+.rg-auth-subtext a {
+  color: rgb(127, 72, 169);
+  text-decoration: none;
+}
+
+.rg-auth-subtext {
+  font-size: 14px;
+  color: #888686;
+}
+
+.rg-auth-form .rg-auth-banner {
+  max-width: 50%;
+  margin: 0 auto;
+}
+
+.rg-auth-form {
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  padding: 20px;
+  text-align: center;
+  align-items: center;
+  min-height: 100vh;
+}
+
+.rg-auth-form .rg-auth-btn {
+  border-radius: 30px;
+  box-shadow: 3px 3px 6x #d3d3d3;
+  color: white;
+  font-weight: bold;
+  background-color: #ffa845;
+  padding: 8px 30px;
+  margin: 10px 0;
+  width: fit-content;
+  margin: 0 auto;
+  transition: 0.2s ease-in-out;
+}
+
+.rg-auth-form .rg-auth-btn:hover {
+  box-shadow: 0px 0px 10px #ffa845;
+}
+
+.rg-auth-form .rg-auth-btn:active {
+  box-shadow: none !important;
+  background-color: white;
+  border: 1px solid #d3d3d3;
+  color: black;
+}
+
+.rg-auth-form .rg-auth-form-inputs {
+  max-width: 600px;
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.rg-auth-form .rg-auth-error {
+  color: red;
+}
+
+.rg-auth-form .rg-auth-input {
+  padding: 8px 20px;
+  border: solid 1px #d3d3d3;
+  border-radius: 30px;
+  margin: 10px 0;
+  width: 100%;
+  transition: 0.5s;
+}
+
+@media(min-width: 400px) {
+  .rg-auth-form {
+      flex-direction: row;
+      justify-content: center;
+      align-items: center;
+  }
+
+  .rg-auth-form .rg-auth-b {
+      max-width: 50%;
+  }
+  
+ .rg-auth-form .rg-auth-form-inputs {
+      min-width: 50%;
+      text-align: center;
+      display: flex;
+      justify-content: center;
+      flex-direction: column;
+  }
+
+  .rg-auth-form .rg-auth-form-inputs  .rg-auth-input {
+      max-width: 400px;
+      margin: 5px auto;
+  }
+}
