--- front-end/src/components/LogIn/LogIn.css
@@ -1,120 +1,0 @@
-.rg-auth-subtext a {
-    color: rgb(127, 72, 169);
-    text-decoration: none;
-}
-
-.rg-auth-subtext {
-    font-size: 14px;
-    color: #888686;
-}
-
-.rg-auth-form .rg-auth-banner {
-    max-height: 50%;
-    margin: 0 auto;
-}
-
-.rg-auth-form {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    padding: 20px;
-    text-align: center;
-    align-items: center;
-    min-height: 100vh;
-}
-
-
-.rg-auth-form .rg-auth-btn {
-    border-radius: 30px;
-    box-shadow: 3px 3px 6x #d3d3d3;
-    color: white;
-    font-weight: bold;
-    background-color: #ffa845;
-    padding: 8px 30px;
-    margin: 10px 0;
-    width: fit-content;
-    margin: 0 auto;
-    transition: 0.2s ease-in-out;
-}
-
-.rg-auth-form .rg-auth-btn:hover {
-    box-shadow: 0px 0px 10px #ffa845;
-}
-
-.rg-auth-form .rg-auth-btn:active {
-    box-shadow: none !important;
-    background-color: white;
-    border: 1px solid #d3d3d3;
-    color: black;
-}
-
-.rg-auth-form .rg-auth-form-inputs {
-    max-width: 600px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-}
-
-.rg-auth-form .rg-auth-input {
-    padding: 8px 20px;
-    border: solid 1px #d3d3d3;
-    border-radius: 30px;
-    margin: 10px 0;
-    width: 100%;
-    transition: 0.5s;
-}
-
-@media(min-width: 400px) {
-    .LogInPage .rg-auth-form {
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-    }
-
-    .LogInPage .rg-auth-form .rg-auth-b {
-        max-width: 50%;
-    }
-    
-   .LogInPage .rg-auth-form .rg-auth-form-inputs {
-        min-width: 50%;
-        text-align: center;
-        display: flex;
-        justify-content: center;
-        flex-direction: column;
-    }
-
-    .LogInPage .rg-auth-form .rg-auth-form-inputs  .rg-auth-input {
-        max-width: 400px;
-        margin: 5px auto;
-    }
-}
-
-.LogInFormContainer{
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: center;
-    width: 100vw;
-    height: 100vh;
-}
-
-.LogInForm{
-    width: 420px;
-    box-shadow: rgb(0 0 0 / 16%) 1px 1px 10px;
-    padding-top: 30px;
-    padding-bottom: 20px;
-    border-radius: 8px;
-    background-color: white;
-    text-align: center;
-}
-
-.usernameInput{
-    padding-top: 20px;
-    padding-bottom: 20px;
-}
-
-.passwordInput{
-    padding-top: 20px;
-    padding-bottom: 20px;
-}
