+++ front-end/src/components/Navbar/Navbar.css
@@ -1,5 +1,5 @@
 .rg-navbar {
-    position: absolute;
+    position: fixed;
     width: 100%;
     display: flex;
     /* justify-content: center; */
